+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 - Enclose Markdown links in anchor tags to make them clickable. Thansk for
   preliminary work done in #85 by @FliegendeWurst.
 
+### Fixed
+
+- Content disposition header value was wrongly constructed.
+
 
 ## 3.0.0